requirements.txt
llama-cpp-python
langchain
faiss-cpu
llama-index
groq
transformers 
sentence-transformers
accelerate
pillow 